requirements.txt
opencv-python
tensorflow
keras
deepface
tf-keras
Pillow 